renamed tab classes, clashing with mod info dialog split nexus tab
